requirements.txt
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7
google-genai==0.3.0
python-dotenv
